requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
boto3==1.28.70
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0